Stash view refactoring
