+++ pages/Data_Import.py
@@ -1,11 +1,12 @@
-from datetime import datetime
-from io import BytesIO
 from threading import Lock
 from typing import List
 
 import pandas as pd
 import streamlit as st
 
+from Home import INITIAL_SIDEBAR_STATE, LAYOUT, MENU_ITEMS, PAGE_ICON
+from widgets.console import show_console
+from widgets.download import download_dataframe_button
 from src.eurostat import (
     cast_time_to_datetimeindex,
     fetch_dataset_and_metadata,
@@ -13,6 +14,16 @@ from src.eurostat import (
     filter_dataset,
     split_dimensions_and_attributes_from,
 )
+
+
+def page_config():
+    st.set_page_config(
+        page_title="Eurostat Data Wizard • Data Import",
+        page_icon=PAGE_ICON,
+        layout=LAYOUT,
+        initial_sidebar_state=INITIAL_SIDEBAR_STATE,
+        menu_items=MENU_ITEMS,  # type: ignore
+    )
 
 
 @st.experimental_singleton(show_spinner=False)
@@ -39,65 +50,8 @@ def load_dataset(code: str) -> pd.DataFrame:
     return data[["flag", "value"]]
 
 
-# `max_entries=1` because likely just the last call will be the reused one
-@st.experimental_memo(show_spinner=False, max_entries=1)
-def load_stash(stash: dict) -> pd.DataFrame:
-    data = pd.DataFrame()
-    common_cols = ["geo", "time", "flag", "value"]
-    for code, filters in stash.items():
-        indexes, flags = filters["indexes"], filters["flags"]
-        df = load_dataset(code)
-        df = filter_dataset(df, indexes, flags)
-        # Append dataset code to data as first level
-        df = pd.concat(
-            {code: df},
-            names=["dataset"],
-        )
-        # Merge dataset-specific indexes into one field
-        df = df.reset_index()
-        df = (
-            pd.concat(
-                [
-                    df[["dataset"]],
-                    df[df.columns.difference(["dataset"] + common_cols)].agg(
-                        " - ".join, axis=1
-                    ),
-                    df[df.columns.intersection(["unit"] + common_cols)],
-                ],
-                axis=1,
-            )
-            .rename(columns={0: "variable"})
-            .set_index(["dataset", "variable"] + common_cols[:-2])
-        )
-        # Append previous loop datasets
-        data = pd.concat([data, df])
-    return data
-
-
 def update_stash(code, indexes, flags):
     st.session_state.stash.update({code: {"indexes": indexes, "flags": flags}})
-
-
-def clear_stash():
-    st.session_state.stash = {}
-
-
-def page_config():
-    st.set_page_config(
-        page_title="Eurostat Data Wizard • Data Import",
-        page_icon="🇪🇺",
-        layout="wide",
-        initial_sidebar_state="expanded",
-        menu_items={
-            "About": """
-            Copyright (c) 2022 Presidenza del Consiglio dei Ministri
-            Datasets are provided [free of charge](https://ec.europa.eu/eurostat/en/about-us/policies/copyright) by © European Union, 1995 - today
-            """,
-        },
-    )
-
-    if "stash" not in st.session_state:
-        st.session_state.stash = {}
 
 
 def patch_streamlit_session_state():
@@ -175,57 +129,24 @@ def import_dataset():
 
 
 def show_dataset(dataset, dataset_code, indexes, flags):
-    st.subheader("Current dataset")
+    st.subheader("Dataset")
     # Dataset is shown with `.reset_index` because MultiIndex are not rendered properly
     view = dataset if dataset.empty else dataset.reset_index()
     st.dataframe(view, use_container_width=True)
     st.write("{} rows x {} columns".format(*view.shape))
 
-    st.button(
-        "Stash",
-        on_click=update_stash,
-        args=(dataset_code, indexes, flags),
-        disabled=dataset.empty,
-    )
-
-
-def show_stash():
-    st.subheader("Stashed datasets")
-
-    dataset = pd.DataFrame()
-    try:
-        with st.spinner(text="Fetching data"):
-            if st.session_state.stash.keys():
-                dataset = load_stash(st.session_state.stash)
-    except ValueError as ve:
-        st.error(ve)
-
-    # Dataset is shown with `.reset_index` because MultiIndex are not rendered properly
-    view = dataset if dataset.empty else dataset.reset_index()
-    st.dataframe(view, use_container_width=True)
-    st.write("{} rows x {} columns".format(*dataset.shape))
-
     with st.container():
         col1, col2 = st.columns(2, gap="small")
         with col1:
-            st.button("Clear", on_click=clear_stash, disabled=dataset.empty)
+            st.button(
+                "Stash",
+                on_click=update_stash,
+                args=(dataset_code, indexes, flags),
+                disabled=dataset.empty,
+            )
         with col2:
-            now = datetime.now().isoformat(timespec="seconds")
-            with BytesIO() as buffer:
-                # Data downloaded is the `view` to be consistent with what user sees
-                view.to_csv(buffer, index=False, compression={"method": "gzip"})
-                st.download_button(
-                    "Download",
-                    buffer.getvalue(),
-                    file_name=f"EurostatDataWizard_{now}.csv.gz",
-                    mime="application/gzip",
-                    disabled=view.empty,
-                )
+            download_dataframe_button(view)
 
-
-def show_console():
-    with st.expander("Session console"):
-        st.write(st.session_state)
 
 
 if __name__ == "__main__":
@@ -233,8 +154,6 @@ if __name__ == "__main__":
 
     dataset, dataset_code, indexes, flags = import_dataset()
 
-    st.header("Data viewer")
     show_dataset(dataset, dataset_code, indexes, flags)
-    show_stash()
 
     show_console()  # For debugging
