+++ widgets/console.py
@@ -1,0 +1,6 @@
+import streamlit as st
+
+
+def show_console():
+    with st.expander("Session console"):
+        st.write(st.session_state)
